requirements.txt
django-cors-headers
boto3
django-storages[s3]
gunicorn==21.2.0
whitenoise==6.6.0
argcomplete==3.6.3
asgiref==3.11.0
attrs==25.4.0
colorama==0.4.6
colorlog==6.10.1
dependency-groups==1.3.1
distlib==0.4.0
Django==5.2.7
djangorestframework==3.16.1
drf-spectacular==0.28.0
djangorestframework-simplejwt==5.5.1
Faker==38.2.0
filelock==3.20.0
inflection==0.5.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
nox==2025.5.1
packaging==25.0
pillow==12.0.0
platformdirs==4.5.0
psycopg2-binary==2.9.9
python-dotenv==1.2.1
PyYAML==6.0.3
referencing==0.37.0
rpds-py==0.29.0
sqlparse==0.5.3
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
virtualenv==20.35.4
